remove normalize_imm

+++ tracer/src/instruction/lb.rs
@@ -82,7 +82,7 @@ impl LB {
             operands: FormatI {
                 rd: v_address,
                 rs1: self.operands.rs1,
-                imm: self.operands.imm as u32 as u64,
+                imm: self.operands.imm as u64,
             },
             virtual_sequence_remaining: Some(7),
         };
@@ -93,7 +93,7 @@ impl LB {
             operands: FormatI {
                 rd: v_word_address,
                 rs1: v_address,
-                imm: -4i64 as u32 as u64,
+                imm: -4i64 as u64,
             },
             virtual_sequence_remaining: Some(6),
         };
