+++ tracer/src/instruction/lhu.rs
@@ -91,7 +91,7 @@ impl LHU {
             operands: FormatI {
                 rd: v_address,
                 rs1: self.operands.rs1,
-                imm: self.operands.imm as u32 as u64,
+                imm: self.operands.imm as u64,
             },
             virtual_sequence_remaining: Some(7),
         };
@@ -102,7 +102,7 @@ impl LHU {
             operands: FormatI {
                 rd: v_word_address,
                 rs1: v_address,
-                imm: -4i64 as u32 as u64,
+                imm: -4i64 as u64,
             },
             virtual_sequence_remaining: Some(6),
         };
