+++ tracer/src/instruction/sb.rs
@@ -84,7 +84,7 @@ impl SB {
             operands: FormatI {
                 rd: v_address,
                 rs1: self.operands.rs1,
-                imm: self.operands.imm as u32 as u64,
+                imm: self.operands.imm as u64,
             },
             virtual_sequence_remaining: Some(12),
         };
@@ -95,7 +95,7 @@ impl SB {
             operands: FormatI {
                 rd: v_word_address,
                 rs1: v_address,
-                imm: -4i64 as u32 as u64,
+                imm: -4i64 as u64,
             },
             virtual_sequence_remaining: Some(11),
         };
@@ -218,7 +218,7 @@ impl SB {
             operands: FormatI {
                 rd: v_address,
                 rs1: self.operands.rs1,
-                imm: self.operands.imm as u32 as u64,
+                imm: self.operands.imm as u64,
             },
             virtual_sequence_remaining: Some(12),
         };
@@ -229,7 +229,7 @@ impl SB {
             operands: FormatI {
                 rd: v_dword_address,
                 rs1: v_address,
-                imm: -8i64 as u32 as u64,
+                imm: -8i64 as u64,
             },
             virtual_sequence_remaining: Some(11),
         };
