+++ tracer/src/instruction/sh.rs
@@ -100,7 +100,7 @@ impl SH {
             operands: FormatI {
                 rd: v_address,
                 rs1: self.operands.rs1,
-                imm: self.operands.imm as u32 as u64,
+                imm: self.operands.imm as u64,
             },
             virtual_sequence_remaining: Some(12),
         };
@@ -111,7 +111,7 @@ impl SH {
             operands: FormatI {
                 rd: v_word_address,
                 rs1: v_address,
-                imm: -4i64 as u32 as u64,
+                imm: -4i64 as u64,
             },
             virtual_sequence_remaining: Some(11),
         };
@@ -244,7 +244,7 @@ impl SH {
             operands: FormatI {
                 rd: v_address,
                 rs1: self.operands.rs1,
-                imm: self.operands.imm as u32 as u64,
+                imm: self.operands.imm as u64,
             },
             virtual_sequence_remaining: Some(12),
         };
@@ -255,7 +255,7 @@ impl SH {
             operands: FormatI {
                 rd: v_dword_address,
                 rs1: v_address,
-                imm: -8i64 as u32 as u64,
+                imm: -8i64 as u64,
             },
             virtual_sequence_remaining: Some(11),
         };
